requirements.txt
urllib3<2.1
requests>=2.31.0
beautifulsoup4==4.12.3
portalocker==2.8.0
python-dotenv>=1.1.0

# --- AWS & Cloud ---
boto3>=1.34.50
awscli>=1.32.50
botocore>=1.34.50
moto==5.0.0
fastmcp>=2.0.0
pydantic>=2.0.0
duckduckgo-search>=5.0.0

# --- AI Orchestration ---
langchain-core>=0.1.0
langchain>=0.1.0
langgraph>=0.0.1
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
langchain-ollama>=0.1.0

# --- UI & TUI (Rich/Textual) ---
rich>=13.9.4
textual==0.52.1

# --- Infrastructure & Automation ---
docker>=7.1.0
cookiecutter==2.6.0

# --- Testing & Quality ---
pytest==8.0.2
pytest-cov==4.1.0
tftest==1.8.0
coverage==7.4.3
